Move core EMLE model functionality to EMLEBase class
