+++ emle/models/__init__.py
@@ -25,6 +25,7 @@
 # avoid severe module import overheads when running the client code,
 # which requires no EMLE functionality.
 
+from ._emle_base import EMLEBase
 from ._emle import EMLE
 from ._ani import ANI2xEMLE
 from ._mace import MACEEMLE
